fix hash seeding
